Explode metadata.json into individual files
The GitHub PR resource type by jtarchie [1] created files that
contained the PR information [2].  Each file contained one piece of
information (e.g., the git branch name or the author).  This format
was convenient since many resource types are configured by passing the
name of a file that contains the option's value.  For example, the
`tag_file` option of the Docker Image resource type accepts the name
of a file whose content is the desired tag [3].

The telia-oss GitHub resource type puts all the PR information in a
single file, metadata.json.  To extract data from this file, it's
necessary to create a new task.

This commit modifies `in.go` so that each individual PR information is
also written into its own file.  For example, `.git/resource/head_name`
contains the name of the git branch.  It helps make some tasks -- such
as using the git branch name as the Docker tag -- as simple as they
were when using the jtarchie resource type.

[1]: https://github.com/jtarchie/github-pullrequest-resource
[2]: https://github.com/jtarchie/github-pullrequest-resource#additional-files-populated
[3]: https://github.com/concourse/docker-image-resource#out-push-an-image-or-build-and-push-a-dockerfile