Fix for new resource version created when one file matches ignore and the other does not
When paths and ignore_paths are both defined, and the pull request
contains one file in the ignored path and another outside of paths
and outside ignore_paths, a new resource version is created because
the number of files matching ignore_paths does not match the number
of files changed. This behaviour is incorrect because there may be
changed files that do not match paths and do not match ignore_paths.